test: split parts-export filter tests from full watertight export for CI shard balance
partsExport.test.ts ran 84s as one file. The part-name filter and
not-found diagnostics tests move to partsExport.filters.test.ts; the
full watertight export stays. No dedup possible: each test calls
runAndExportParts with a different parts option. Test counts unchanged:
3 -> 1+2. Measured per-file: 24.4s + 25.7s.
